+++ anki-common/src/main/res/values/sentence-case.xml
